+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetitemop.py
@@ -380,5 +380,5 @@ class CreateDatasetItemResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2025-03-04T21:45:15.768Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2025-03-04T22:54:38.560Z")
     r"""The date and time the resource was last updated"""