+++ packages/orq-rc/src/orq_ai_sdk/models/updatepromptop.py
@@ -517,8 +517,8 @@ class UpdatePromptRequestBodyTypedDict(TypedDict):
     domain_id: NotRequired[str]
     created: NotRequired[str]
     updated: NotRequired[str]
-    created_by_id: NotRequired[str]
-    updated_by_id: NotRequired[str]
+    created_by_id: NotRequired[Nullable[str]]
+    updated_by_id: NotRequired[Nullable[str]]
     display_name: NotRequired[str]
     r"""The prompt’s name, meant to be displayable in the UI."""
     description: NotRequired[Nullable[str]]
@@ -537,9 +537,9 @@ class UpdatePromptRequestBody(BaseModel):
 
     updated: Optional[str] = None
 
-    created_by_id: Optional[str] = None
+    created_by_id: OptionalNullable[str] = UNSET
 
-    updated_by_id: Optional[str] = None
+    updated_by_id: OptionalNullable[str] = UNSET
 
     display_name: Optional[str] = None
     r"""The prompt’s name, meant to be displayable in the UI."""
@@ -566,7 +566,7 @@ class UpdatePromptRequestBody(BaseModel):
             "prompt_config",
             "metadata",
         ]
-        nullable_fields = ["description"]
+        nullable_fields = ["created_by_id", "updated_by_id", "description"]
         null_default_fields = []
 
         serialized = handler(self)
@@ -1197,8 +1197,8 @@ class UpdatePromptResponseBodyTypedDict(TypedDict):
     r"""The prompt’s name, meant to be displayable in the UI."""
     prompt_config: UpdatePromptPromptConfigTypedDict
     r"""A list of messages compatible with the openAI schema"""
-    created_by_id: NotRequired[str]
-    updated_by_id: NotRequired[str]
+    created_by_id: NotRequired[Nullable[str]]
+    updated_by_id: NotRequired[Nullable[str]]
     description: NotRequired[Nullable[str]]
     r"""The prompt’s description, meant to be displayable in the UI. Use this field to optionally store a long form explanation of the prompt for your own purpose"""
     metadata: NotRequired[UpdatePromptPromptsMetadataTypedDict]
@@ -1225,9 +1225,9 @@ class UpdatePromptResponseBody(BaseModel):
     prompt_config: UpdatePromptPromptConfig
     r"""A list of messages compatible with the openAI schema"""
 
-    created_by_id: Optional[str] = None
+    created_by_id: OptionalNullable[str] = UNSET
 
-    updated_by_id: Optional[str] = None
+    updated_by_id: OptionalNullable[str] = UNSET
 
     description: OptionalNullable[str] = UNSET
     r"""The prompt’s description, meant to be displayable in the UI. Use this field to optionally store a long form explanation of the prompt for your own purpose"""
@@ -1237,7 +1237,7 @@ class UpdatePromptResponseBody(BaseModel):
     @model_serializer(mode="wrap")
     def serialize_model(self, handler):
         optional_fields = ["created_by_id", "updated_by_id", "description", "metadata"]
-        nullable_fields = ["description"]
+        nullable_fields = ["created_by_id", "updated_by_id", "description"]
         null_default_fields = []
 
         serialized = handler(self)